# Change the time measurement method for final model fitting