@topepo 2022-08-07
Thanks for the contribution and sorry for the long wait. 

`system.time()` is generally preferred but I don't think it is a big deal to use `proc.time()`. I'll update and merge 👍